requirements.txt
aiofiles==23.1.0
aiogram==3.0.0
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
bleach==6.0.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.3
defusedxml==0.7.1
Django==4.2.4
django-bootstrap5==23.3
django-filter==23.2
django-summernote==0.8.20.0
django-taggit==4.0.0
djangorestframework==3.14.0
drf-spectacular==0.26.4
frozenlist==1.4.0
idna==3.4
inflection==0.5.1
jsonschema==4.18.6
jsonschema-specifications==2023.7.1
magic-filter==1.0.11
Markdown==3.4.4
multidict==6.0.4
oauthlib==3.2.2
Pillow==10.0.0
psycopg2==2.9.6
pycparser==2.21
pydantic==2.3.0
pydantic_core==2.6.3
PyJWT==2.8.0
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0.1
redis==4.6.0
referencing==0.30.1
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.9.2
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==2.0.4
webencodings==0.5.1
yarl==1.9.2
